Reviewed version of LRTDP
The main difference with the previous version is that a new feauture to deal with cyclic graphs has been added.
I was supposed to work at the "checkSolved" level but this method already implemented a mechanism to avoid overhead due to cycles. This method doesn't append an already checked state to the openstack again.

By contrast, I've implemented  a filter that removes the possibility of duplicated instances of a state in the "visited" list. This means that the back-propagation and the labeling precedure, which are done in reverse order of the trace ( from the last state to the beginning), could encounter gaps. For instance: if the trace is like s0-s1-s0-s4-s5-s6-s7-s11  the states are appended s0-s1-s4-s5-s6-s7-s11. This doesn't seem to have a major effect on the convergence. I would be interesting to study if the saved computational time is  big enough to support this modification.